// src/models/userModel.ts
export interface IUserState {
    user_id: number | null;
    unique_id: number | null;
    name: string | null;
    mobileNo: string | null;
    image: string | null;
    online_status: string | null;
    country_code: string | null;
    country_id: number | null;
    currency: string | null;
    apiStatus: string | null;
    logs: any;
}